+++ app/utopia/routes/airport_controller.py
@@ -2,10 +2,15 @@ from flask import Flask, json, request, make_response, render_template, jsonify
 from utopia.models.flights import AIRPORT_SCHEMA_MANY, ROUTE_SCHEMA_MANY
 from utopia import app
 from utopia.service.airport_service import AirportService
+from utopia.models.users import find_user, refresh_token
 import logging
+from flask_jwt_extended import get_current_user, JWTManager, jwt_required
+
+
 
 AIRLINE_SERVICE = AirportService()
-
+jwt = JWTManager(app)
+ADMIN = 1
 
 
 ################### GET ###################
@@ -28,7 +33,7 @@ def readAirports():
 @app.route('/airlines/read/routes', methods=['GET'])
 def readRoutes():
 
-    return AIRLINE_SERVICE.read_routes(1) 
+    return AIRLINE_SERVICE.read_routes() 
 
 
 @app.route('/airlines/find/airport/id=<iata_id>', methods=['GET'])
@@ -53,19 +58,31 @@ def readRoutesByAirport(direction, iata_id):
 
 
 @app.route('/airlines/add/airport', methods=['POST'])
+@jwt_required()
 def addAirport():
+    current_user = get_current_user()
+    if current_user['role_id'] != ADMIN:
+        return make_response('need admin priveleges to access this resource', 403)
 
     return AIRLINE_SERVICE.create_airport(request.json)
 
 
 @app.route('/airlines/add/airport/test', methods=['POST'])
+@jwt_required()
 def addAirportTest():
+    current_user = get_current_user()
+    if current_user['role_id'] != ADMIN:
+        return make_response('need admin priveleges to access this resource', 403)
 
     AIRLINE_SERVICE.add_airports_test(request.json)
     return ''
 
 @app.route('/airlines/add/route', methods=['POST'])
+@jwt_required()
 def addRoute():
+    current_user = get_current_user()
+    if current_user['role_id'] != ADMIN:
+        return make_response('need admin priveleges to access this resource', 403)
 
     return AIRLINE_SERVICE.add_route(request.json)
 
@@ -74,13 +91,21 @@ def addRoute():
 
 
 @app.route('/airlines/update/airport', methods=['PUT'])
+@jwt_required()
 def updateAirport():
+    current_user = get_current_user()
+    if current_user['role_id'] != ADMIN:
+        return make_response('need admin priveleges to access this resource', 403)
 
     return AIRLINE_SERVICE.update_airport(request.json)
 
 
 @app.route('/airlines/update/route', methods=['PUT'])
+@jwt_required()
 def updateRoute():
+    current_user = get_current_user()
+    if current_user['role_id'] != ADMIN:
+        return make_response('need admin priveleges to access this resource', 403)
 
     return AIRLINE_SERVICE.update_route(request.json)
 
@@ -89,15 +114,32 @@ def updateRoute():
 
 
 @app.route('/airlines/delete/airport/id=<iata_id>', methods=['DELETE'])
+@jwt_required()
 def deleteAirport(iata_id):
+    current_user = get_current_user()
+    if current_user['role_id'] != ADMIN:
+        return make_response('need admin priveleges to access this resource', 403)
 
     AIRLINE_SERVICE.delete_airport(iata_id)
     return ''
 
 
 @app.route('/airlines/delete/route/id=<id>', methods=['DELETE'])
+@jwt_required()
 def deleteRoute(id):
+    current_user = get_current_user()
+    if current_user['role_id'] != ADMIN:
+        return make_response('need admin priveleges to access this resource', 403)
 
     return AIRLINE_SERVICE.delete_route(id)
+
+@jwt.user_lookup_loader
+def user_lookup_callback(_jwt_header, jwt_data):
+    identity = jwt_data["sub"]
+    return find_user(identity)
+
+@app.after_request
+def refresh_expiring_jwts(response):
+    return refresh_token(response)
